Created a delete save file button in main menu
Created a delete save file button in the options panel only in the main menu
We do not want the player to be able to delete the save file of the game they are currently playing on from the pause menu.
I also cleared some of the white space in the SaveSystemMk2.cs
I also modified the Load() function in there to check if the file exists first, if not then just load new game.